Refactor scenario generation
